requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
Brotli==1.0.9
celery==5.3.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
Django==4.2.4
django-appconf==1.0.5
django-crispy-forms==2.0
django-imagekit==4.1.0
django-ordered-model==3.7.4
django-paypal==2.0
et-xmlfile==1.1.0
fonttools==4.42.1
gunicorn==21.2.0
html5lib==1.1
idna==3.4
kombu==5.3.2
openpyxl==3.1.2
packaging==23.1
pilkit==2.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
